+++ ots/source/creature.h
@@ -7,7 +7,7 @@
 // modify it under the terms of the GNU General Public License
 // as published by the Free Software Foundation; either version 2
 // of the License, or (at your option) any later version.
-// 
+//
 // This program is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
@@ -35,7 +35,7 @@
 extern LuaScript g_config;
 
 #ifdef YUR_HIGH_LEVELS
-typedef __int64 exp_t;
+typedef int64_t exp_t;
 #else
 typedef unsigned long exp_t;
 #endif //YUR_HIGH_LEVELS
@@ -107,14 +107,14 @@ public:
 		if(condIt != this->end() && !condIt->second.empty()) {
 			return true;
 		}
-		
+
 		return false;
 	}
 };
 
 
 //////////////////////////////////////////////////////////////////////
-// Defines the Base class for all creatures and base functions which 
+// Defines the Base class for all creatures and base functions which
 // every creature has
 
 class Creature : public AutoID, public Thing
@@ -123,16 +123,16 @@ public:
 	//Creature(const std::string& name);
 	Creature();
 	virtual ~Creature();
-	
+
 	virtual const std::string& getName() const = 0;
-	
+
 	void setID(){this->id = auto_id | this->idRange();}
 	virtual unsigned long idRange() = 0;
 	unsigned long getID() const { return id; }
 	virtual void removeList() = 0;
 	virtual void addList() = 0;
 
-	exp_t getExpForLv(const int& lv) const { 
+	exp_t getExpForLv(const int& lv) const {
 #ifdef YUR_HIGH_LEVELS
 		exp_t x = lv;
 		return ((50*x/3 - 100)*x + 850/3)*x - 200;
@@ -143,14 +143,14 @@ public:
 
 	Direction getDirection() const { return direction; }
 	void setDirection(Direction dir) { direction = dir; }
-	
+
 	virtual fight_t getFightType(){return FIGHT_MELEE;};
 	virtual subfight_t getSubFightType() {return DIST_NONE;}
 	virtual Item* getDistItem() {return NULL;};
 	virtual void removeDistItem(){return;}
 	virtual int getImmunities() const
 	{
-		if(access >= g_config.ACCESS_PROTECT) 
+		if(access >= g_config.ACCESS_PROTECT)
 			return  ATTACK_ENERGY | ATTACK_BURST | ATTACK_FIRE |
 			ATTACK_PHYSICAL | ATTACK_POISON | ATTACK_PARALYZE | ATTACK_DRUNKNESS;
 		else
@@ -158,24 +158,24 @@ public:
 	};
 
 #ifdef YUR_PVP_ARENA
-	virtual void drainHealth(int, CreatureVector* arenaLosers);
+	virtual void drainHealth(int64_t, CreatureVector* arenaLosers);
 #else
 	virtual void drainHealth(int);
 #endif //YUR_PVP_ARENA
 
-	virtual void drainMana(int);
+	virtual void drainMana(int64_t);
 	virtual void die(){};
 	virtual std::string getDescription(bool self = false) const;
 	virtual void setAttackedCreature(const Creature* creature);
 	//virtual void setAttackedCreature(unsigned long id);
-	
+
 	virtual void setMaster(Creature* creature);
 	virtual Creature* getMaster() {return master;}
-	
+
 	virtual void addSummon(Creature *creature);
 	virtual void removeSummon(Creature *creature);
-	
-	virtual int getWeaponDamage() const {
+
+	virtual int64_t getWeaponDamage() const {
 		return 1+(int)(10.0*rand()/(RAND_MAX+1.0));
 	}
 	virtual int getArmor() const {
@@ -184,33 +184,33 @@ public:
 	virtual int getDefense() const {
 		return 0;
 	}
-	
+
 	unsigned long attackedCreature;
-	
+
 	virtual bool isAttackable() const { return true; };
 	virtual bool isPushable() const {return true;}
 	virtual void dropLoot(Container *corpse) {return;};
 	virtual int getLookCorpse() {return lookcorpse;};
-	
+
 	//  virtual int sendInventory(){return 0;};
 	virtual int addItemInventory(Item* item, int pos){return 0;};
 	virtual Item* getItem(int pos){return NULL;}
 	virtual Direction getDirection(){return direction;}
 	void addCondition(const CreatureCondition& condition, bool refresh);
 	Conditions& getConditions() {return conditions;};
-	
+
 	int lookhead, lookbody, looklegs, lookfeet, looktype, lookcorpse, lookmaster;
-	int mana, manamax, manaspent;
+	int64_t mana, manamax, manaspent;
 	bool pzLocked;
-	
+
 	long inFightTicks, exhaustedTicks;
 	long manaShieldTicks, hasteTicks, paralyzeTicks;
 	int immunities;
-	
+
 	//unsigned long experience;
 	Position masterPos;
-	
-	int health, healthmax;
+
+	int64_t health, healthmax;
 	uint64_t lastmove;
 
 #ifdef TJ_MONSTER_BLOOD
@@ -218,55 +218,55 @@ public:
     unsigned char bloodeffect;
     unsigned char bloodsplash;
 #endif //TJ_MONSTER_BLOOD
-	
+
 	long long getSleepTicks() const;
 	int getStepDuration() const;
-	
-	unsigned short getSpeed() const {            
+
+	unsigned short getSpeed() const {
 		return speed;
 	};
-	
+
 	void setNormalSpeed()
 	{
 		if(access >= g_config.ACCESS_PROTECT){
-			speed = 1000;     
-			return;    
+			speed = 1000;
+			return;
 		}
 		speed = getNormalSpeed();
 	}
 
 	int getNormalSpeed()
 	{
-		if(access >= g_config.ACCESS_PROTECT){     
-			return 1000;    
+		if(access >= g_config.ACCESS_PROTECT){
+			return 1000;
 		}
 #ifdef YUR_BOH
 	#ifdef YUR_RINGS_AMULETS
-		return std::min(900, 220 + (2* (level + 30*boh + 30*timeRing - 1))); 
+		return std::min(900, 220 + (2* (level + 30*boh + 30*timeRing - 1)));
 	#else //YUR_RINGS_AMULETS
-		return std::min(900, 220 + (2* (level + 30*boh - 1))); 
+		return std::min(900, 220 + (2* (level + 30*boh - 1)));
 	#endif //YUR_RINGS_AMULETS
 #else //YUR_BOH
 	#ifdef YUR_RINGS_AMULETS
-		return std::min(900, 220 + (2* (level + 30*timeRing - 1))); 
+		return std::min(900, 220 + (2* (level + 30*timeRing - 1)));
 	#else //YUR_RINGS_AMULETS
-		return std::min(900, 220 + (2* (level - 1))); 
+		return std::min(900, 220 + (2* (level - 1)));
 	#endif //YUR_RINGS_AMULETS
 #endif //YUR_BOH
 	}
-	
+
 	int access;		//access level
-	int maglevel;	// magic level
+	int64_t maglevel;	// magic level
 	int level;		// level
 	int speed;
-	
+
 	Direction direction;
-	
+
 	virtual bool canMovedTo(const Tile *tile) const;
-	
+
 	virtual void sendCancel(const char *msg) const { };
 	virtual void sendCancelWalk(const char *msg) const { };
-	
+
 	virtual void addInflictedDamage(Creature* attacker, int damage);
 	virtual exp_t getGainedExperience(Creature* attacker);
 	virtual std::vector<long> getInflicatedDamageCreatureList();
@@ -287,7 +287,7 @@ public:
 	void setInvisible(int ticks) { invisibleTicks = ticks; }
 	bool checkInvisible(int thinkTicks);
 #endif //YUR_INVISIBLE
-	
+
 protected:
 
 #ifdef YUR_BOH
@@ -305,17 +305,17 @@ protected:
 
 	Creature *master;
 	std::list<Creature*> summons;
-	
+
 	Conditions conditions;
 	typedef std::vector< std::pair<uint64_t, long> > DamageList;
 	typedef std::map<long, DamageList > TotalDamageList;
 	TotalDamageList totaldamagelist;
-	
+
 protected:
 	virtual int onThink(int& newThinkTicks){newThinkTicks = 300; return 300;};
 	virtual void onThingMove(const Creature *player, const Thing *thing, const Position *oldPos,
 		unsigned char oldstackpos, unsigned char oldcount, unsigned char count) { };
-	
+
 	virtual void onCreatureAppear(const Creature *creature) { };
 	virtual void onCreatureDisappear(const Creature *creature, unsigned char stackPos, bool tele = false) { };
 	virtual void onThingDisappear(const Thing* thing, unsigned char stackPos) = 0;
@@ -323,50 +323,50 @@ protected:
 	virtual void onThingAppear(const Thing* thing) = 0;
 	virtual void onCreatureTurn(const Creature *creature, unsigned char stackPos) { };
 	virtual void onCreatureSay(const Creature *creature, SpeakClasses type, const std::string &text) { };
-	
+
 	virtual void onCreatureChangeOutfit(const Creature* creature) { };
 	virtual void onTileUpdated(const Position &pos) { };
-	
+
 	virtual void onTeleport(const Creature *creature, const Position *oldPos, unsigned char oldstackpos) { };
-	
+
 	//container to container
 	virtual void onThingMove(const Creature *creature, const Container *fromContainer, unsigned char from_slotid,
 		const Item* fromItem, int oldFromCount, Container *toContainer, unsigned char to_slotid,
 		const Item *toItem, int oldToCount, int count) {};
-	
+
 	//inventory to container
 	virtual void onThingMove(const Creature *creature, slots_t fromSlot, const Item* fromItem,
 		int oldFromCount, const Container *toContainer, unsigned char to_slotid, const Item *toItem, int oldToCount, int count) {};
-	
+
 	//inventory to inventory
 	virtual void onThingMove(const Creature *creature, slots_t fromSlot, const Item* fromItem,
 		int oldFromCount, slots_t toSlot, const Item* toItem, int oldToCount, int count) {};
-	
+
 	//container to inventory
 	virtual void onThingMove(const Creature *creature, const Container *fromContainer, unsigned char from_slotid,
 		const Item* fromItem, int oldFromCount, slots_t toSlot, const Item *toItem, int oldToCount, int count) {};
-	
+
 	//container to ground
 	virtual void onThingMove(const Creature *creature, const Container *fromContainer, unsigned char from_slotid,
 		const Item* fromItem, int oldFromCount, const Position &toPos, const Item *toItem, int oldToCount, int count) {};
-	
+
 	//inventory to ground
 	virtual void onThingMove(const Creature *creature, slots_t fromSlot,
 		const Item* fromItem, int oldFromCount, const Position &toPos, const Item *toItem, int oldToCount, int count) {};
-	
+
 	//ground to container
 	virtual void onThingMove(const Creature *creature, const Position &fromPos, int stackpos, const Item* fromItem,
 		int oldFromCount, const Container *toContainer, unsigned char to_slotid, const Item *toItem, int oldToCount, int count) {};
-	
+
 	//ground to inventory
 	virtual void onThingMove(const Creature *creature, const Position &fromPos, int stackpos, const Item* fromItem,
 		int oldFromCount, slots_t toSlot, const Item *toItem, int oldToCount, int count) {};
-	
+
 	friend class Game;
 	friend class Map;
 	friend class Commands;
 	friend class GameState;
-	
+
 	unsigned long id;
 	//std::string name;
 };
